backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djoser==2.1.0
django-filter==23.3
djangorestframework-csv==3.0.1
gunicorn==20.1.0
Pillow==10.1.0
python-dotenv==1.0.1